fix: migrations TINYINT → SMALLINT (PostgreSQL não suporta TINYINT)

+++ apps/api/src/migrations/1700000000133-DbConnectionIsPrincipal.ts
@@ -4,7 +4,7 @@ export class DbConnectionIsPrincipal1700000000133 implements MigrationInterface 
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(`
       ALTER TABLE db_connections
-      ADD COLUMN IF NOT EXISTS isPrincipal TINYINT(1) NOT NULL DEFAULT 0
+      ADD COLUMN IF NOT EXISTS isPrincipal SMALLINT NOT NULL DEFAULT 0
     `);
   }
